+++ Demo/Sources/iOS/Pulse_Demo_iOSApp.swift
@@ -25,8 +25,15 @@ private final class AppViewModel: ObservableObject {
     init() {
         // This code registers the store with the `RemoteLogger` (important!)
         LoggerStore.shared = .demo
+        
+        /*
+            //Disable: options -> 1. "Settings", "Get Pulse Pro", "Report Issue"
+        UserDefaults.standard.set(true, forKey: "pulse-disable-support-prompts")
+        UserDefaults.standard.set(true, forKey: "pulse-disable-report-issue-prompts")
+        UserDefaults.standard.set(true, forKey: "pulse-disable-settings-prompts")
+         */
 
-        RemoteLogger.shared.isAutomaticConnectionEnabled = true
+        // RemoteLogger.shared.isAutomaticConnectionEnabled = true
 
         // NetworkLogger.enableProxy()