+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Pulse 5.x
 
+## Pulse 5.1.2
+
+*Oct 10, 2024*
+
+- Fix an issue with `isAutomaticConnectionEnabled` not starting the browser the first time it's enabled
+- Fix crash when exporting HAR file by @rounak in https://github.com/kean/Pulse/pull/299
+- Fix concurrency issue in `StoreDetailsViewModel` by @ejensen in https://github.com/kean/Pulse/pull/302
+- Update the deep search. It will now show all search scopes found and only one match per scope. You can tap on a match to see the scope with a prepopulated search query to see the remaining items.
+- Fix an issue with deep search not working for request and response headers
+
 ## Pulse 5.1.1
 
 *Sep 19, 2024*